requirements_vlm.txt
# MenuGen VLM Version Dependencies
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.21.0
diffusers>=0.21.0
huggingface_hub>=0.16.0

# Vision/Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Web Framework
streamlit>=1.28.0

# Utilities
numpy>=1.24.0
requests>=2.28.0

# Optional: For better performance
# bitsandbytes>=0.41.0  # For quantization if needed
